image separation and integrated
