+++ modules/image/lang.json
@@ -1,17 +1,18 @@
 {
   "en": {
+    "accept": "Accept ",
+    "accepted": "Accepted ",
     "aki": "AKI - Amazon Kernel Image",
     "ami": "AMI - Amazon Machine Image",
     "architecture": "Architecture ",
     "ari": "ARI - Amazon Ramdisk Image",
+    "add_data": "Add",
     "checksum": "Checksum ",
     "docker": "Docker",
-    "edit_image": "Edit Image ",
-    "edit_status": "Edit Status ",
     "format": "Format",
-    "info": "Information ",
     "image_id": "Image ID",
     "image_name": "Image Name",
+    "info": "Information ",
     "iso": "ISO - Optical Disk Image",
     "meta_data": "Meta Data",
     "min_disk": "Minimum Disk(GB)",
@@ -19,61 +20,72 @@
     "no_architecture": "Not Choose",
     "no_members": "No Available Image To Shared",
     "ova": "OVA - Open Virtual Appliance",
-    "project_id": "Project ID ",
+    "pending": "Pending ",
     "protected": "Protected ",
-    "public": "Public ",
-    "public_image": "Public Image ",
+    "project_id": "Project Id",
     "qcow2": "QCOW2 - QEMU Emulator",
     "raw": "Raw",
-    "related_instance": "Related Instance ",
-    "resource_type": "Resource Type",
+    "reject": "Reject ",
+    "rejected": "Rejected ",
+    "related_instance": "Related Instance",
     "resttime": ", Rest Time: ",
-    "saving": "Saving ",
-    "snapshot_type": "Snapshot",
+    "saving": "Saving",
+    "share": "Share ",
+    "shared_image": "Shared Image ",
     "speed": ", Speed: ",
+    "key_value": "Key",
+    "private": "Private ",
+    "private_image": "Private Image ",
     "upload_canael": "Upload Canceled ",
     "upload_progress": "Upload Progress ",
     "url": "URL ",
+    "value": "Value",
     "vdi": "VDI - Virtual Disk Image",
     "vhd": "VHD - Virtual Hard Disk",
-    "visibility": "Visibility",
+    "visibility": "visibility",
     "vmdk": "VMDK - Virtual Machine Disk"
   },
   "zh-CN": {
+    "accept": "接受",
+    "accepted": "已接受",
+    "add_data": "添加",
     "aki": "AKI - 亚马逊内核镜像",
     "ami": "AMI - Amazon Machine 镜像",
     "architecture": "架构",
     "ari": "ARI - Amazon Ramdisk 镜像",
     "checksum": "校验和",
     "docker": "Docker",
-    "edit_image": "修改镜像",
-    "edit_status": "修改状态",
     "format": "格式",
-    "info": "信息",
     "image_id": "镜像ID",
     "image_name": "镜像名称",
+    "info": "信息",
     "iso": "ISO - 光盘镜像",
+    "key_value": "键值",
     "meta_data": "元数据",
     "min_disk": "最小磁盘(GB)",
     "min_ram": "最低内存(MB)",
     "no_architecture": "无",
-    "no_members": "没有镜像被分享",
+    "no_members": "没有镜像被共享",
     "ova": "OVA - 开源虚拟化应用",
-    "project_id": "项目ID",
+    "pending": "等待中",
+    "private": "私有",
+    "private_image": "私有镜像",
     "protected": "是否受保护",
-    "public": "公有",
-    "public_image": "公有镜像",
+    "project_id": "项目ID",
     "qcow2": "QCOW2 - QEMU 模拟器",
     "raw": "Raw",
+    "reject": "拒绝",
+    "rejected": "已拒绝",
     "related_instance": "关联云主机",
-    "resource_type": "资源类型",
     "resttime": ", 剩余时间: ",
     "saving": "保存中",
-    "snapshot_type": "快照",
+    "share": "共享",
+    "shared_image": "共享镜像",
     "speed": ", 速度: ",
     "upload_canael": "上传已取消",
     "upload_progress": "上传进度",
     "url": "URL",
+    "value": "值",
     "vdi": "VDI - 虚拟磁盘镜像",
     "vhd": "VHD - 虚拟硬盘",
     "visibility": "可见性",
