+++ modules/image/request.js
@@ -1,5 +1,6 @@
 const storage = require('client/applications/dashboard/cores/storage');
 const fetch = require('client/applications/dashboard/cores/fetch');
+let RSVP = require('rsvp');
 
 module.exports = {
   getList: function(forced) {
@@ -12,34 +13,14 @@ module.exports = {
       url: '/api/v1/' + HALO.user.projectId + '/overview'
     });
   },
-  getVolumeTypes: function() {
-    return fetch.get({
-      url: '/proxy/cinder/v2/' + HALO.user.projectId + '/types'
-    });
-  },
-  getVolumePrice: function(type, size) {
-    let url = '/proxy/gringotts/v2/products/price' +
-      '?purchase.bill_method=hour' +
-      '&purchase.purchases[0].product_name=' + type +
-      '&purchase.purchases[0].service=block_storage' +
-      '&purchase.purchases[0].region_id=' + HALO.current_region +
-      '&purchase.purchases[0].quantity=' + size;
-
-    return fetch.get({
-      url: url
-    });
-  },
   getPrices: function() {
     return fetch.get({
       url: '/proxy/gringotts/v2/products'
     });
   },
-  createVolume: function(_data) {
-    let data = {};
-    data.volume = _data;
-
-    return fetch.post({
-      url: '/proxy/cinder/v2/' + HALO.user.projectId + '/volumes',
+  updateImage: function(imageID, data) {
+    return fetch.patch({
+      url: '/api/v1/images/' + imageID,
       data: data
     });
   },
@@ -48,9 +29,9 @@ module.exports = {
       return data.instance;
     });
   },
-  updateImage: function(imageID, data) {
-    return fetch.patch({
-      url: '/api/v1/images/' + imageID,
+  createImage: function(data) {
+    return fetch.post({
+      url: '/proxy/glance/v2/images',
       data: data
     });
   },
@@ -60,15 +41,49 @@ module.exports = {
       data: data
     });
   },
+  getShared: function() {
+    let that = this;
+    return fetch.get({
+      url: '/proxy/glance/v1/shared-images/' + HALO.user.projectId
+    }).then(res => {
+      let sharedImages = res.shared_images,
+        deferredList = [];
+      sharedImages.forEach(image => {
+        deferredList.push(that.getDetail(image.image_id));
+      });
+      return RSVP.all(deferredList);
+    });
+  },
+  getDetail: function(imageId) {
+    return fetch.get({
+      url: '/proxy/glance/v2/images/' + imageId + '/members/' + HALO.user.projectId
+    });
+  },
+  getImageDetail: function(members) {
+    let deferredList = [];
+    deferredList = members.map(member => {
+      return fetch.get({
+        url: '/proxy/glance/v2/images/' + member.image_id
+      });
+    });
+
+    return RSVP.all(deferredList);
+  },
+  updateMember: function(member, data) {
+    return fetch.put({
+      url: '/proxy/glance/v2/images/' + member.id + '/members/' + member.member_id,
+      data: data
+    });
+  },
   deleteImage: function(id) {
     return fetch.delete({
       url: '/proxy/glance/v2/images/' + id
     });
   },
-  createImage: function(data) {
+  createMember: function(imageId, id) {
     return fetch.post({
-      url: '/proxy/glance/v2/images',
-      data: data
+      url: '/proxy/glance/v2/images/' + imageId + '/members',
+      data: {member: id}
     });
   }
 };
