--- modules/private-image/detail/style/index.less
@@ -1,45 +1,0 @@
-@import "../../../../style/theme.less";
-
-@pagi-guide-color: @color_font_800;
-@detail-border-color: @color_grey_600;
-
-.halo-module-related-instance {
-  position: absolute;
-  top: 36px;
-  left: 200px;
-  right: 0;
-  bottom: 0;
-  background-color: #fff;
-  border-left: 1px solid @detail-border-color;
-  .detail-content {
-    height: 100%;
-    display: flex;
-    flex-direction: column;
-    margin: 0 20px;
-    > div {
-      position: relative;
-    }
-
-    .table-with-no-data > .table,
-    > .table {
-      margin-top: 12px;
-      position: relative;
-    }
-    > .pagination-box {
-      .page-guide {
-        padding-right: 10px;
-        color: @pagi-guide-color;
-        vertical-align: middle;
-      }
-      .pagination {
-        vertical-align: middle;
-      }
-      position: relative;
-      flex: 1;
-      height: 60px;
-      min-height: 60px;
-      max-height: 60px;
-      text-align: right;
-    }
-  }
-}
