--- modules/private-image/request.js
@@ -1,74 +1,0 @@
-const storage = require('client/applications/dashboard/cores/storage');
-const fetch = require('client/applications/dashboard/cores/fetch');
-
-module.exports = {
-  getList: function(forced) {
-    return storage.getList(['image', 'instance', 'network', 'keypair', 'flavor'], forced).then(function(data) {
-      return data.image;
-    });
-  },
-  getOverview: function() {
-    return fetch.get({
-      url: '/api/v1/' + HALO.user.projectId + '/overview'
-    });
-  },
-  getVolumeTypes: function() {
-    return fetch.get({
-      url: '/proxy/cinder/v2/' + HALO.user.projectId + '/types'
-    });
-  },
-  getVolumePrice: function(type, size) {
-    let url = '/proxy/gringotts/v2/products/price' +
-      '?purchase.bill_method=hour' +
-      '&purchase.purchases[0].product_name=' + type +
-      '&purchase.purchases[0].service=block_storage' +
-      '&purchase.purchases[0].region_id=' + HALO.current_region +
-      '&purchase.purchases[0].quantity=' + size;
-
-    return fetch.get({
-      url: url
-    });
-  },
-  getPrices: function() {
-    return fetch.get({
-      url: '/proxy/gringotts/v2/products'
-    });
-  },
-  createVolume: function(_data) {
-    let data = {};
-    data.volume = _data;
-
-    return fetch.post({
-      url: '/proxy/cinder/v2/' + HALO.user.projectId + '/volumes',
-      data: data
-    });
-  },
-  getInstances: function() {
-    return storage.getList(['instance'], false).then(data => {
-      return data.instance;
-    });
-  },
-  updateImage: function(imageID, data) {
-    return fetch.patch({
-      url: '/api/v1/images/' + imageID,
-      data: data
-    });
-  },
-  createTask(data) {
-    return fetch.post({
-      url: '/proxy/glance/v2/tasks',
-      data: data
-    });
-  },
-  deleteImage(id) {
-    return fetch.delete({
-      url: '/proxy/glance/v2/images/' + id
-    });
-  },
-  createMember(imageId, id) {
-    return fetch.post({
-      url: '/proxy/glance/v2/images/' + imageId + '/members',
-      data: {member: id}
-    });
-  }
-};
