--- modules/private-image/style/index.less
@@ -1,4 +1,0 @@
-.halo-module-image {
-  flex: 1;
-  display: flex;
-}
