--- modules/shared-image/config.json
@@ -1,88 +1,0 @@
-{
-  "tabs": [{
-    "name": ["public_image"],
-    "key": "image"
-  }, {
-    "name": ["shared_image"],
-    "key": "shared-image",
-    "default": true
-  }, {
-    "name": ["private_image"],
-    "key": "private-image"
-  }],
-  "btns": [{
-    "value": ["create", "instance"],
-    "key": "create",
-    "type": "create",
-    "icon": "create",
-    "disabled": true
-  }, {
-    "value": ["create", "volume"],
-    "key": "create_volume",
-    "type": "create",
-    "icon": "create",
-    "disabled": true
-  }, {
-    "value": ["update", "status"],
-    "key": "update_status",
-    "icon": "edit"
-  }, {
-    "key": "refresh",
-    "icon": "refresh"
-  }],
-  "search": {
-    "column": {
-      "name": true
-    }
-  },
-  "table": {
-    "checkbox": true,
-    "dataKey": "id",
-    "loading": true,
-    "column": [{
-      "title": ["name"],
-      "dataIndex": "name",
-      "key": "name",
-      "type": "captain",
-      "sort": true
-    }, {
-      "title": ["size"],
-      "dataIndex": "size",
-      "key": "size"
-    }, {
-      "title": ["type"],
-      "key": "type",
-      "dataIndex": "image_type"
-    }, {
-      "title": ["status"],
-      "dataIndex": "status",
-      "type": "status",
-      "key": "status",
-      "filter": [{
-        "name": ["active"],
-        "key": "active",
-        "filterBy": "active"
-      }]
-    }, {
-      "title": ["create", "time"],
-      "dataIndex": "created_at",
-      "key": "created",
-      "type": "time"
-    }, {
-      "title": ["visibility"],
-      "dataIndex": "visibility",
-      "key": "visibility"
-    }],
-    "data": [],
-    "detail": {
-      "tabs": [{
-        "name": ["description"],
-        "key": "description",
-        "default": true
-      }, {
-        "name": ["related_instance"],
-        "key": "instance"
-      }]
-    }
-  }
-}
