--- modules/shared-image/lang.json
@@ -1,20 +1,0 @@
-{
-  "en": {
-    "accept": "Accept ",
-    "accepted": "Accepted ",
-    "pending": "Pending ",
-    "reject": "Reject ",
-    "rejected": "Rejected ",
-    "share": "Shared ",
-    "shared_image": "Shared Image "
-  },
-  "zh-CN": {
-    "accept": "接受",
-    "accepted": "已接受",
-    "pending": "等待中",
-    "reject": "拒绝",
-    "rejected": "已拒绝",
-    "share": "共享",
-    "shared_image": "共享镜像"
-  }
-}
