--- modules/shared-image/pop/create_volume/config.json
@@ -1,46 +1,0 @@
-{
-  "title": ["create", "volume"],
-  "fields": [{
-    "type": "icon_label",
-    "field": "image",
-    "icon_type": "",
-    "text": ""
-  }, {
-    "type": "input",
-    "field": "name"
-  }, {
-    "type": "tab",
-    "field": "type",
-    "data": [],
-    "value": "",
-    "hide": true
-  }, {
-    "type": "slider",
-    "field": "capacity_size",
-    "min": 1,
-    "max": 1,
-    "value": 1,
-    "unit": "GB",
-    "hide": true
-  }, {
-    "type": "charge",
-    "field": "charge",
-    "has_label": true,
-    "hide": true
-  }, {
-    "type": "tip",
-    "tip_type": "warning",
-    "field": "warning",
-    "hide": true
-  }, {
-    "type": "tip",
-    "tip_type": "error",
-    "field": "error",
-    "hide": true
-  }],
-  "btn": {
-    "value": "create",
-    "type": "create",
-    "disabled": true
-  }
-}
